[feat] 댓글 완성 최최최종본

+++ board_view.js
@@ -46,6 +46,7 @@ docsd.forEach((eachdoc) => {
   let num = row["num"];
   let when = row["when"];
   let commentNum = row["commentNum"];
+  let commentPw = row["commentPw"];
   let id = eachdoc.id;
   let which;
 
@@ -60,8 +61,6 @@ docsd.forEach((eachdoc) => {
       <div class="resultBottom">
       <div id="resultTime">${when}</div>
       <div id="DeleteBtn">
-      <button class="" id="commentUpdate">수정
-      <button style="display : none">${commentNum}</button>
       <button class="" id="commentDelete">삭제
       <button style="display : none">${commentNum}</button>
       </div>
@@ -71,24 +70,24 @@ docsd.forEach((eachdoc) => {
     $("#result").append(append_comment);
   }
   // 댓글 삭제
-  $("#commentField").click(async function (e) {
+  $('#commentField').click(async function (e) {
     e.preventDefault();
     let clickCoNum = e.target.nextElementSibling.innerText;
-
-    // console.log('clickCoNum => ', clickCoNum)
-    // console.log('commentNum =>', commentNum)
+    // 
+    console.log('clickCoNum => ', clickCoNum)
+    console.log('commentNum =>', commentNum)
     if (clickCoNum === commentNum) {
-      await deleteDoc(doc(db, 'comments', which));
-      window.location.reload();
+      const pw = prompt("삭제 비밀번호를 입력해주세요.");
+      if (pw === commentPw) {
+        await deleteDoc(doc(db, 'comments', which));
+        window.location.reload();
+      } else {
+        alert("비밀번호가 올바르지 않습니다..")
+      }
     } else {
       return false;
     }
 
-
-    // if (id === which) {
-    //   await deleteDoc(doc(db, 'comments', which));
-    //   window.location.reload();
-    // }
   })
 })
 
@@ -184,20 +183,20 @@ $("#commentBtn").click(async function (e) {
     commentPw: $("#pwd1").val()
 
   };
-// 댓글 폼 규칙
+  // 댓글 폼 규칙
   if (data.commentName.length <= 0 || data.commentText.length <= 0) {
     return alert('내용을 입력해주세요.');
-  }else if ($('#pwd1').val().length <= 0 || $('#pwd2').val().length <= 0) {
+  } else if ($('#pwd1').val().length <= 0 || $('#pwd2').val().length <= 0) {
     return alert('비밀번호를 입력해주세요.')
-  }else if ($('#pwd1').val() != $('#pwd2').val()) {
+  } else if ($('#pwd1').val() != $('#pwd2').val()) {
     return alert('비밀번호를 일치하도록 입력해주세요.')
-  }else {
+  } else {
     await addDoc(collection(db, "comments"), data);
     console.log(data)
     window.location.reload()
   }
 })
-  
+
 // 댓글 비밀번호 일치 확인     
 $(function () {
   $('#pwd1').keyup(function () {
